Move to o.a.b.util.http.HttpToolResponse - https://github.com/apache/incubator-brooklyn/pull/994, and
